add updates for mosaicking
