+++ skysat_stereo/skysat_stereo_workflow.py
@@ -12,9 +12,10 @@ from skysat_stereo import asp_utils as asp
 from rpcm import geo
 from skysat_stereo import misc_geospatial as misc
 from shapely.geometry import Polygon
-from itertools import combinations,compress
+import itertools
 from osgeo import osr
 from pyproj import Transformer
+
 
 def prepare_stereopair_list(img_folder,perc_overlap,out_fn,aoi_bbox=None,cross_track=False):
     """
@@ -29,7 +30,7 @@ def prepare_stereopair_list(img_folder,perc_overlap,out_fn,aoi_bbox=None,cross_t
         sys.exit()
     out_shp = os.path.splitext(out_fn)[0]+'_bound.gpkg'
     n_proc = iolib.cpu_count()
-    shp_list = p_umap(skysat.skysat_footprint,img_list,num_cpus=2*n_proc)
+    shp_list = p_map(skysat.skysat_footprint,img_list,num_cpus=2*n_proc)
     merged_shape = misc.shp_merger(shp_list)
     bbox = merged_shape.total_bounds
     merged_shape = misc.shp_merger(shp_list)
@@ -56,7 +57,7 @@ def prepare_stereopair_list(img_folder,perc_overlap,out_fn,aoi_bbox=None,cross_t
         mask = merged_shape.to_crs(bbox.crs).intersects(bbox)
         img_list = merged_shape[mask].img.values
 
-    img_combinations = list(combinations(img_list,2))
+    img_combinations = list(itertools.combinations(img_list,2))
     n_comb = len(img_combinations)
     perc_overlap = np.ones(n_comb,dtype=float)*perc_overlap
     proj = local_aea
@@ -64,8 +65,8 @@ def prepare_stereopair_list(img_folder,perc_overlap,out_fn,aoi_bbox=None,cross_t
     # result to this contains truth value (0 or 1, overlap percentage)
     truth_value = [tvs[0] for tvs in tv]
     overlap = [tvs[1] for tvs in tv]
-    valid_list = list(compress(img_combinations,truth_value))
-    overlap_perc_list = list(compress(overlap,truth_value))
+    valid_list = list(itertools.compress(img_combinations,truth_value))
+    overlap_perc_list = list(itertools.compress(overlap,truth_value))
     print('Number of valid combinations are {}, out of total {}  input images making total combinations {}\n'.format(len(valid_list),len(img_list),n_comb))
     with open(out_fn, 'w') as f:
         img1_list = [x[0] for x in valid_list]
@@ -97,7 +98,6 @@ def skysat_preprocess(img_folder,mode,sampling=None,frame_index=None,product_lev
         except:
             os.makedirs(outdir)
     if mode == 'video':
-        sampling = args.video_sampling_mode
         frame_index = skysat.parse_frame_index(frame_index,True)
         product_level = 'l1a'
         num_samples = len(frame_index)
@@ -469,3 +469,115 @@ def align_cameras_wrapper(input_camera_list,transform_txt,outfolder,rpc=0,dem='N
     p_umap(asp.align_cameras,input_camera_list,transform_list,outfolder,write,rpc,dem,
            img_list,num_cpus = iolib.cpu_count())
     
+
+def dem_mosaic_wrapper(dir,mode='triplet',out_folder=None,identifier=None,tile_size=None,filter_dem=1,min_video_count=2,max_video_nmad=5):
+    if out_folder is None:
+        out_folder = os.path.join(dir,'composite_dems')
+    
+    if not os.path.exists(out_folder):
+        os.makedirs(out_folder)
+    if identifier is None:
+        identifier = ''
+    if mode == 'triplet':
+        dir_list = sorted(glob.glob(os.path.join(dir,'20*/')))
+        print(f"Number of combinations {len(dir_list)}")
+        
+        def valid_disp(direc):
+            try:
+                D_sub = iolib.fn_getma(os.path.join(direc,'run-D_sub.tif'),3)
+                stats = np.percentile(D_sub.compressed(),(2,98))
+                direc_out = glob.glob(os.path.join(direc,'run*{}*-DEM.tif'.format(identifier)))[0]
+                out = (direc_out,True)
+            except:
+                out = (0,False)
+            return out
+
+    
+        # find all valid DEMs
+        total_cpu = iolib.cpu_count()
+        n_proc = total_cpu - np.arange(len(dir_list))
+        #this setup is so that old pools are discarded and new ones with new number of workers are created
+        valid_dem_dir_list = []
+        for idx,direc in enumerate(tqdm(dir_list)):
+            comb_dir_list = sorted(glob.glob(os.path.join(direc,'*/')))
+            results = p_map(valid_disp,comb_dir_list,num_cpus=n_proc[idx])
+            t_val = [r[1] for r in results]
+            valid_dirs = list(itertools.compress([r[0] for r in results],t_val))
+            valid_dem_dir_list.append(valid_dirs)
+        
+        # naming logic for pairwise and triplet/multiview composites
+        mdt1,t1,mdt2,t2 = [[],[],[],[]]
+        combination_out_list = []
+        if len(dir_list)>3:
+            print("Input is cross-track")
+        if dir_list[0][-1] == '/':
+            dir_list = [x[:-1] for x in dir_list]
+        for direc in dir_list:
+            combination_out_list.append(os.path.join(out_folder,os.path.basename(direc)+'_wt_avg_mos.tif'))
+            a,b,c,d = os.path.basename(direc).split('_')
+            mdt1.append(a) #master date (year month date)
+            t1.append(b) # time of day in seconds
+            mdt2.append(c)
+            t2.append(d)
+        if len(direc)>3:
+            composite_prefix = 'multiview_'+'_'.join(np.unique(mdt1+mdt2))+'__'+'_'.join(np.unique(t1+t2))
+        else:
+            composite_prefix = 'triplet_'+'_'.join(np.unique(mdt1+mdt2))+'__'+'_'.join(np.unique(t1+t2))
+            
+        # produce bistereo pairwise mosaics
+        len_combinations = len(combination_out_list)
+        
+        
+        if len_combinations > 3:
+            # force tiled processing in case of multiview mosaicking
+            if not tile_size:
+                tile_size = 400
+        mos_log = p_map(asp.dem_mosaic,valid_dem_dir_list,combination_out_list,
+                ['None']*len_combinations,[None]*len_combinations,
+                [None]*len_combinations,[tile_size]*len_combinations)
+
+        if len_combinations > 2:
+            print("Producing triplet/multiview composites")
+            total_dem_list = list(itertools.chain.from_iterable(valid_dem_dir_list))
+            print(f"Mosaicing {len(total_dem_list)} DEM strips using median, wt_avg, count, nmad operators")
+            stats_list = [None,'median','nmad','count']
+            out_fn_list = [os.path.join(out_folder,
+                            '{}_{}_mos.tif'.format(composite_prefix,stat)) for stat in ['wt_avg','median','nmad','count']]
+            composite_mos_log = p_map(asp.dem_mosaic,[total_dem_list]*4,out_fn_list,['None']*4,[None]*4,stats_list,
+                                [tile_size]*4,num_cpus=4)
+        out_log_fn = os.path.join(out_folder,'skysat_triplet_dem_mos.log')
+        print("Saving triplet DEM mosaic log at {}".format(out_log_fn))
+        with open(out_log_fn,'w') as f:
+            for log in mos_log+composite_mos_log:
+                f.write(log) 
+    if mode == 'video':
+        dir_list = sorted(glob.glob(os.path.join(dir,'1*/')))
+        valid_video_dir = []
+        for video_dir in dir_list:
+            try:
+                D_sub = iolib.fn_getma(os.path.join(video_dir,'run-D_sub.tif'),3)
+                stats = [np.percentile(D_sub.compressed(),(2,98)),np.mean(D_sub.compressed())]
+                DEM = glob.glob(os.path.join(video_dir,'run*{}*-DEM.tif'.format(identifier)))[0]
+                valid_video_dir.append(video_dir)
+            except:
+                continue 
+        video_dem_list = [glob.glob(os.path.join(dir,f'run*{identifier}*-DEM.tif'))[0] for dir in valid_video_dir]
+        stats_list = ['median','count','nmad']
+        print('total dems are {}'.format(len(video_dem_list)))
+        out_fn_list = [os.path.join(out_folder,'video_{}_mos.tif'.format(stat)) for stat in stats_list]
+        dem_mos_log = p_map(asp.dem_mosaic,[video_dem_list]*3,out_fn_list,['None']*3,[None]*3,stats_list,[None]*3) 
+        out_log_fn = os.path.join(out_folder,'skysat_video_dem_mos.log')
+        with open(out_log_fn,'w') as f:
+            for log in dem_mos_log:
+                f.write(log)
+        if filter_dem == 1:
+            print("Filtering DEM using NMAD and count metrics")
+            print(f"Filter will use min count of {min_video_count} and max NMAD of {max_video_nmad}")
+            mos_ds_list = warplib.memwarp_multi_fn(out_fn_list)
+            # Filtered array list contains dem_filtered,nmad_filtered, count_filtered in order
+            filtered_array_list = skysat.filter_video_dem_by_nmad(mos_ds_list,min_video_count,max_video_nmad)
+            trailing_str = f'_filt_max_nmad{max_video_nmad}_min_count{min_video_count}.tif'
+            out_filter_fn_list = [os.path.splitext(fn)[0]+trailing_str for fn in out_fn_list]
+            for idx,fn in enumerate(out_filter_fn_list):
+                iolib.writeGTiff(filtered_array_list[idx],fn,mos_ds_list[idx])
+
